% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping.R
\name{ping}
\alias{ping}
\title{Flatten resource}
\usage{
ping(host = "localhost", port = "8000")
}
\arguments{
\item{host}{path to a datapackage.json file}

\item{port}{base resource to flatten}
}
\description{
Ping the flask web application.
}
\examples{
\dontrun{
ping()

ping(host = "ckan")
}
}
